updating experimental RAG intrinsics to granite 3.2 (#138)
We need to escape the JSON string in the prompt so langchain does
not get upset when it formats the prompt as an f-string to insert
the query and the documents.

Since the response from the model for hallucinations and citations is
text, we do not want to be attempting to parse that text since it
cannot be API.

Finally since this replaces a recipe that used ollama and no other
recipes use ollama, we remove the ollama part of the CI build.

Signed-off-by: Kelson Martins <kelson.martins@proton.me>
Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>
Co-authored-by: BJ Hargrave <hargrave@us.ibm.com>