Update rag recipes for granite 3.3's new document role
Granite 3.3 no longer uses a single 'documents' role to hold all the rag
documents. We must now use a 'document' role per document. Since
langchain builds the documents part of the prompt in the rag chain, we
update the document prompt template to include the special granite
tokens and format for the document role.

This is not the best solution but it appears to work with minimal
changes.

Signed-off-by: BJ Hargrave <hargrave@us.ibm.com>
Signed-off-by: BJ Hargrave <bj@hargrave.dev>